Move next_inputs tests out of operator.rs into operator_test.rs.
Same sibling layout as chained_operator_test. Tests go through OperatorBase::next_inputs.
