+++ src/runtime/operators/mod.rs
@@ -4,6 +4,8 @@ pub mod sink;
 pub mod source;
 pub mod key_by;
 pub mod operator;
+#[cfg(test)]
+mod operator_test;
 pub mod kind;
 pub mod chained;
 pub mod aggregate;
